+++ src/models/Settings.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace samuelreichoer\queryapi\models;
+
+use craft\base\Model;
+
+class Settings extends Model
+{
+    public ?int $cacheDuration = null;
+}